+++ libvast/vast/system/terminate.hpp
@@ -38,36 +38,40 @@ namespace vast::system {
 /// Performs an asynchronous shutdown of a set of actors by sending an EXIT
 /// message, configurable either in sequential or parallel mode of operation.
 /// As soon as all actors have terminated, the returned promise gets fulfilled.
+/// This function is the lower-level interface for bringing down actors. The
+/// function `shutdown` uses this function internally to implement a more
+/// convenient one-stop solution.
 /// @param self The actor to terminate.
 /// @param xs The actors to terminate.
+/// @param grace_period The amount of time to wait until all actors terminated
+///        cleanly.
+/// @param kill_timeout The timeout befor giving and delivering an error to the
+///        response promise.
 /// @returns A response promise to be fulfilled when all *xs* terminated.
-template <class Policy>
-auto terminate(caf::event_based_actor* self, std::vector<caf::actor> xs) {
-  auto timeout = defaults::system::shutdown_timeout;
-  return self->request(self->spawn(terminator<Policy>), timeout, std::move(xs));
-}
-
-template <class Policy, class... Ts>
-auto terminate(caf::typed_event_based_actor<Ts...>* self,
-               std::vector<caf::actor> xs) {
-  auto handle = caf::actor_cast<caf::event_based_actor*>(self);
-  return terminate<Policy>(handle, std::move(xs));
-}
-
-template <class Policy>
-void terminate(caf::scoped_actor& self, std::vector<caf::actor> xs) {
-  auto timeout = defaults::system::shutdown_timeout;
-  self->request(self->spawn(terminator<Policy>), timeout, std::move(xs))
-    .receive(
-      [=](atom::done) { VAST_DEBUG_ANON("terminated all dependent actor"); },
-      [=](const caf::error& err) {
-        VAST_ERROR_ANON("failed to terminated all dependent actors", err);
-      });
+/// @relates shutdown
+template <class Policy, class Actor>
+auto terminate(Actor&& self, std::vector<caf::actor> xs,
+               std::chrono::milliseconds grace_period
+               = defaults::system::shutdown_grace_period,
+               std::chrono::milliseconds kill_timeout
+               = defaults::system::shutdown_kill_timeout) {
+  auto t = self->spawn(terminator<Policy>, grace_period, kill_timeout);
+  auto request_timeout = grace_period + kill_timeout;
+  if (request_timeout > std::chrono::milliseconds::zero())
+    return self->request(std::move(t), request_timeout, std::move(xs));
+  else
+    return self->request(std::move(t), caf::infinite, std::move(xs));
 }
 
 template <class Policy, class Actor>
-auto terminate(Actor* self, caf::actor x) {
-  return terminate<Policy>(self, std::vector<caf::actor>{std::move(x)});
+auto terminate(Actor&& self, caf::actor x,
+               std::chrono::milliseconds grace_period
+               = defaults::system::shutdown_grace_period,
+               std::chrono::milliseconds kill_timeout
+               = defaults::system::shutdown_kill_timeout) {
+  return terminate<Policy>(std::forward<Actor>(self),
+                           std::vector<caf::actor>{std::move(x)}, grace_period,
+                           kill_timeout);
 }
 
 } // namespace vast::system