+++ CHANGELOG.md
@@ -11,6 +11,12 @@ Every entry has a category for which we use the following visual abbreviations:
 
 ## Unreleased
 
+- 🐞 The shutdown process of the server process could potentially hang forever.
+  VAST now uses a 2-step procedure that first attempts to terminate all
+  components cleanly. If that fails, it will attempt a hard kill afterwards,
+  and if that fails after another timeout, the process will call `abort(3)`.
+  [#1005](https://github.com/tenzir/vast/pull/1005)
+
 - 🐞 Some file descriptors remained open when they weren't needed any more.
   This descriptor leak has been fixed.
   [#1018](https://github.com/tenzir/vast/pull/1018)